+++ Source/Generate/Test.wren
@@ -1,9 +1,7 @@
-import "soup" for Soup
+import "soup" for Soup, SoupExtension
 
-class TestClass {
-	static runBefore { [] }
-
-	static doStuff() {
+class TestClass is SoupExtension {
+	static evaluate() {
 		Soup.debug("I am running in a method!")
 		Soup.warning("I am running in a method!")
 		Soup.error("I am running in a method!")
